Got perspective projection working correctly!  Currently have three cubes chillin.  Added meshes vector and vectors for names and centers, in order to keep track of position of centers of meshes throughout operation.

+++ Engine/Math/vector3.cpp
@@ -61,7 +61,7 @@ vector3 vector3::Project(vector3 a, vector3 b)
 
 float vector3::Magnitude(vector3 in)
 {
-   float mag = sqrt((in.x * in.x) + (in.y * in.y) + (in.z * in.z));
+   float mag = (float)sqrt((in.x * in.x) + (in.y * in.y) + (in.z * in.z));
    return mag;
 }
